requirements.txt
geomloss==0.2.6
matplotlib==3.7.0
numpy==1.24.2
PyYAML==6.0.2
scipy==1.14.1
torch==2.7.1
tqdm==4.66.1
